Crate pallet_aliases

This pallet aliases allows users to create human readable aliases to address their owned existing assets. Example:

A user owns a Proto that represents an image of “batman” on chain. This asset is addressable using its hash, which is not human readable and impossible to remember. This pallet-aliases allows the user to create an alias like “DC/batman” that links to that hash and can be used by the owner to demonstrate ownership without the need to use the hash of the asset.

“DC/batman” –> f9480f9ead9b82690fodf65546kjmyg730f12763ca2f50ce1792

An alias is composed of two parts: /. The namespace is a unique string that identifies the owner. The alias is a string that identifies the asset.

The user must be owner of the linked asset in order to create the alias.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

LinkTargetVersioned

Struct that add versioning to a LinkTarget using block number as version number.

Enums

LinkTarget

Enum that indicates the types of target assets linked to an alias